backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["siesto1elemento <rohitojha9720@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
python-multipart = "^0.0.20"
faiss-cpu = "^1.11.0"
openparse = "^0.7.0"
dotenv = "^0.9.9"
openai = "^1.86.0"
black = "^25.1.0"
websockets = "^15.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
